+++ packages/glance_widget_platform_interface/lib/glance_widget_platform_interface.dart
@@ -3,6 +3,7 @@ library;
 export 'src/glance_widget_platform.dart';
 export 'src/method_channel_glance_widget.dart';
 export 'src/types/glance_exception.dart';
+export 'src/types/glance_widget_update.dart';
 export 'src/types/widget_action.dart';
 export 'src/types/widget_data.dart';
 export 'src/types/widget_theme.dart';